+++ test/Sentry.Profiling.Tests/SamplingTransactionProfilerTests.cs
@@ -227,7 +227,7 @@ public class SamplingTransactionProfilerTests
         // Disable process exit flush to resolve "There is no currently active test." errors.
         options.DisableAppDomainProcessExitFlush();
 
-        options.AddIntegration(new ProfilingIntegration(TimeSpan.FromSeconds(10)));
+        options.AddProfilingIntegration(TimeSpan.FromSeconds(10));
 
         try
         {
@@ -318,7 +318,7 @@ public class SamplingTransactionProfilerTests
             ProfilesSampleRate = 1.0,
         };
 
-        options.AddIntegration(new ProfilingIntegration(TimeSpan.FromSeconds(10)));
+        options.AddProfilingIntegration(TimeSpan.FromSeconds(10));
 
         try
         {
@@ -365,7 +365,7 @@ public class SamplingTransactionProfilerTests
             TracesSampleRate = 1.0,
             ProfilesSampleRate = 0,
         };
-        options.AddIntegration(new ProfilingIntegration());
+        options.AddProfilingIntegration();
         using var hub = new Hub(options);
         Assert.Null(hub.Options.TransactionProfilerFactory);
     }
@@ -379,7 +379,7 @@ public class SamplingTransactionProfilerTests
             TracesSampleRate = 0,
             ProfilesSampleRate = 1.0,
         };
-        options.AddIntegration(new ProfilingIntegration());
+        options.AddProfilingIntegration();
         using var hub = new Hub(options);
         Assert.Null(hub.Options.TransactionProfilerFactory);
     }
@@ -393,7 +393,7 @@ public class SamplingTransactionProfilerTests
             TracesSampleRate = 1.0,
             ProfilesSampleRate = 1.0,
         };
-        options.AddIntegration(new ProfilingIntegration());
+        options.AddProfilingIntegration();
         using var hub = new Hub(options);
         Assert.NotNull(hub.Options.TransactionProfilerFactory);
     }